requirements.txt
asgiref==3.6.0
black==22.12.0
boto3==1.26.50
botocore==1.29.50
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
cssselect2==0.7.0
dj-rest-auth==2.2.6
Django==4.2
django-cors-headers==3.13.0
django-mathfilters==1.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.38.0
gunicorn==20.1.0
html5lib==1.1
imgkit==1.2.3
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tomli==2.0.1
urllib3==1.26.14
weasyprint==57.2
webencodings==0.5.1
zopfli==0.2.2
